refactor: better folder structure
