+++ src/database/DatabasePersistance/CustomersDatabasePersistance/__tests__/helpers/createCustomersDatabase.ts
@@ -1,4 +1,4 @@
-import DatabaseConnection from "../../../DatabaseConnection/DatabaseConnection";
+import DatabaseConnection from "../../../../DatabaseConnection/DatabaseConnection";
 import CustomersDatabasePersistance from "../../CustomersDatabasePersistance";
 
 export default async function createCustomersDatabase() {
