+++ src/database/DatabasePersistance/CustomersDatabasePersistance/declaration/CustomerRecord.ts
